feat(web): fetcher functions intro
